Move Documentation to the `@0xproject/react-docs` package
