+++ packages/react-docs/.npmignore
@@ -1,0 +1,5 @@
+.*
+yarn-error.log
+/src/
+/scripts/
+tsconfig.json
