+++ packages/react-docs/CHANGELOG.md
@@ -1,0 +1,3 @@
+# CHANGELOG
+
+## vX.X.X - _TBD, 2018_
