+++ packages/react-docs/src/ts/components/enum.tsx
@@ -1,8 +1,9 @@
 import * as _ from 'lodash';
 import * as React from 'react';
-import { EnumValue } from 'ts/types';
 
-interface EnumProps {
+import { EnumValue } from '../types';
+
+export interface EnumProps {
     values: EnumValue[];
 }
