+++ packages/react-shared/src/ts/globals.d.ts
@@ -1,1 +1,7 @@
 declare module 'react-highlight';
+
+// is-mobile declarations
+declare function isMobile(): boolean;
+declare module 'is-mobile' {
+    export = isMobile;
+}
