+++ packages/react-shared/src/ts/index.ts
@@ -5,7 +5,7 @@ export { MarkdownSection } from './components/markdown_section';
 export { NestedSidebarMenu } from './components/nested_sidebar_menu';
 export { SectionHeader } from './components/section_header';
 
-export { HeaderSizes, Styles, MenuSubsectionsBySection } from './types';
+export { HeaderSizes, Styles, MenuSubsectionsBySection, EtherscanLinkSuffixes, Networks } from './types';
 
 export { utils } from './utils/utils';
 export { constants } from './utils/constants';
