+++ packages/website/package.json
@@ -18,6 +18,7 @@
     "author": "Fabio Berger",
     "license": "Apache-2.0",
     "dependencies": {
+        "@0xproject/react-docs": "^0.0.1",
         "@0xproject/react-shared": "^0.0.1",
         "@0xproject/subproviders": "^0.6.0",
         "@0xproject/utils": "^0.4.0",
@@ -25,14 +26,12 @@
         "accounting": "^0.4.1",
         "basscss": "^8.0.3",
         "blockies": "^0.0.2",
-        "compare-versions": "^3.0.1",
         "dateformat": "^2.0.0",
         "deep-equal": "^1.0.1",
         "dharma-loan-frame": "^0.0.12",
         "ethereumjs-tx": "^1.3.3",
         "ethereumjs-util": "^5.1.1",
         "find-versions": "^2.0.0",
-        "is-mobile": "^0.2.2",
         "jsonschema": "^1.2.0",
         "less": "^2.7.2",
         "lodash": "^4.17.4",
